[OG-138] Code review
