+++ polymorphia-frontend/components/course-groups/modals/index.css
@@ -20,3 +20,7 @@
     @apply max-w-44 truncate;
   }
 }
+
+.course-group-modal-loading {
+  @apply w-88 h-[220px] flex-col-centered relative;
+}
