+++ polymorphia-frontend/components/course-groups/modals/move-animal/index.tsx
@@ -42,7 +42,7 @@ function MoveAnimalModalContent() {
 
   if (isLoading) {
     return (
-      <div className="w-88 h-[220px] flex-col-centered relative">
+      <div className="course-group-modal-loading">
         <Loading />
       </div>
     );
